Adv 311 claimable rewards in the form of ibc tokens are not displayed on staking portal (#52)
* modify the UI to show token rewards

* Modify the logic to add token rewards in the staking position

Still missing ticker and decimals for the time being.

* Update types.ts

Added commment and modified reward type to match API response

* Updated helper.ts

Retrieve information from API response.

* Re-add comment and add tokenId from Reward type

* replace tokenId with token.id

* Ensure all token info are present to show a token reward

By excluding tokens with incomplete data, we ensure that the user is not presented with potentially incorrect or confusing information.

Before processing each reward.token, the code checks that all necessary fields (id, ticker, name, and decimals) are available. If any of these fields are missing, the token is skipped.

* align with main

* Revert "align with main"

This reverts commit ea0a2f75b8bb3f2f4c19bb5be1bdd92f400f081e.

* remove TokenAmount value

* update stakingposition types to use TokenAmount

* Update helper.ts to use TokenAmount

* Update StakingPositionsList.tsx

The StakingPositionsListRow component has been updated to use the new TokenAmount[] structure, ensuring that the rewards are displayed correctly. The changes should be compatible with the rest of your application as long as TokenAmount[] is used consistently across your codebase.

* show 0.00001> for small value

* Improve display of IBC tokens when very low values

	•	Refactor StakingPositionsList to handle small token values.
	•	Group and display tokens with values < 0.00001.
	•	Fix JSX and TypeScript errors for better stability.